pyproject.toml
[project]
name = "mindscopex_analysis"
version = "0.1.0"
description = "MindScopeX: 페르소나·레이어 활성화·SAE·activation patching 연구"
requires-python = ">=3.11,<3.14"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "accelerate>=0.30.0",
    "nnsight>=0.3.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "plotly>=5.22.0",
    "kaleido>=0.2.1",
    "matplotlib>=3.8.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
sae = [
    "sae-lens>=4.0.0",
]
dev = [
    "ipykernel>=6.29.0",
    "jupyter>=1.0.0",
    "ruff>=0.6.0",
]

[project.scripts]
mindscopex-run = "mindscopex_analysis.cli:main"

[build-system]
requires = ["hatchling>=1.21.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mindscopex_analysis"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
